+++ index.js
@@ -2,6 +2,7 @@
 import express from 'express';
 import morgan from 'morgan';
 import users from './server/routes/user';
+import mentors from './server/routes/mentor';
 
 
 const app = express();
@@ -10,6 +11,7 @@ app.use(morgan('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use('/api/v1/', users);
+app.use('/api/v1/', mentors);
 
 
 const port = process.env.PORT || 3000;
