+++ server/routes/user.js
@@ -1,14 +1,10 @@
 import express from 'express';
-import UsersController from '../controllers/user';
-import checkAuth from '../middleware/checkAuth';
-import checkAdmin from '../middleware/checkAdmin';
+import { usersSignUp, usersSignIn } from '../controllers/user';
 import signUpValidator from '../middleware/signUpValidator';
 import signinValidator from '../middleware/signInValidator';
 
 const router = express.Router();
-router.post('/auth/signup', signUpValidator, UsersController.usersSignUp);
-router.post('/auth/signIn', signinValidator, UsersController.usersSignIn);
-router.patch('/user/:userId', [checkAuth, checkAdmin], UsersController.userChangeToMentor);
-router.get('/mentors', checkAuth, UsersController.allMentors);
+router.post('/auth/signup', signUpValidator, usersSignUp);
+router.post('/auth/signIn', signinValidator, usersSignIn);
 
-module.exports = router;
+export default router;
